fix: Move get_all_commands to naff.ext.jurigged
